+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix tracking new items when the new-item shine is disabled.
 * Added option to Farming Mode to not move weapons and armor to make space for engrams.
 * About and Support pages are now translatable.
+* Improved error handling and error messages.
 
 # 3.13.0